Add pytest to CI (#3)
This is NOT textbook pytest. If that were the focus, you should use `src/` and `tests/` as specified here [Pytest: Tests outside application code](https://docs.pytest.org/en/7.1.x/explanation/goodpractices.html#tests-outside-application-code).

But this requires a lot of extra overhead with init and project files that would be distracting.